Fix the usage of retrieve/list mixins.
The RetrieveModelMixin and ListModelMixin are usually used in ViewSets.
Here, it's being used with a regular View and so GET requests for
collection (i.e. those without an ID in the path) and resource (i.e.
those with an ID in the path) URL's are both getting dispatched to the
get() method. This causes a problem for openapi spec generation (at
least using drf-spectacular) because it can't distinguish truthfully
whether the view should use a serializer where many=True or not. This
results in the response schema not being accurate for the resource URL's
because drf_spectacular thinks it's a list view since it inherits from
ListModelMixin.

Now, in your urls.py, you can use ChunkedUploadCollectionView.as_view()
for the collection URL and ChunkedUploadResourceView.as_view for the
resource URL. If you need to extend the functionality, like with
on_completed, you can create your own mixin class that implements that
functionality and then new classes that inherit from that and
ChunkedUpload(Collection|Resource)View. Those would then get used in
urls.py.